Small changes to reuse types
