+++ src/nodes/openware-data-historical-merge/openware-data-historical-merge.html/editor.html
@@ -1,18 +1,18 @@
 <script type="text/html" data-template-name="openware-data-historical-merge">
   <div class="form-row">
-    <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-    <input type="text" id="node-input-name" placeholder="Name">
-</div>
-<div class="form-row">
-    <label for="node-input-server"><i class="fa fa-tag"></i> Server</label>
-    <input type="text" id="node-input-server" placeholder="Server...">
-</div>
-<div class="form-row">
-    <label for="node-input-output"><i class="fa fa-file"></i> Output</label>
-    <input type="text" id="node-input-output" placeholder="JSON">
-</div>
-<div class="form-row" id="node-input-delimiter-row">
-    <label for="node-input-delimiter"><i class="fa fa-file"></i> Delimiter</label>
-    <input type="text" id="node-input-delimiter" placeholder=",">
-</div>
+        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
+        <label for="node-input-server"><i class="fa fa-tag"></i> Server</label>
+        <input type="text" id="node-input-server" placeholder="Server...">
+    </div>
+    <div class="form-row">
+        <label for="node-input-output"><i class="fa fa-file"></i> Output</label>
+        <input type="text" id="node-input-output" placeholder="JSON">
+    </div>
+    <div class="form-row" id="node-input-delimiter-row">
+        <label for="node-input-delimiter"><i class="fa fa-file"></i> Delimiter</label>
+        <input type="text" id="node-input-delimiter" placeholder=",">
+    </div>
 </script>
