+++ src/nodes/openware-data-historical-merge/shared/types.ts
@@ -3,16 +3,3 @@ export interface OpenwareDataHistoricalMergeOptions {
   output: "JSON" | "VALUES_ONLY" | "CSV";
   delimiter?: string;
 }
-
-export type PipePayloadType = {
-  pipe: any;
-};
-export type HistoricMergePayloadType = {
-  sensorInfos: {
-    source: string;
-    sensor: string;
-    dimension: number;
-  }[];
-  start: number;
-  end: number;
-};
